WEB-3770 use styled components to abstract out styling
